Fix a potential IP address conflict
Previously the tests generated a random IP address (out of 90)
for each load balancer. As these tests run in parallel there was a
chance that two load balancers could be assigned the same address,
causing a conflict and randomly failing a test.

This patch changes the tests to use a neutron assigned address unless
an explicit "use_fixed_ip" flag is set.
The load balancer create API test uses this flag, the others will use
a neutron assigned address.

Change-Id: Ic22efd1df027fd7dd0db14783f2c76069fd51bb5