Add "wait for active" after object delete
Currently we are waiting for the object to finish deleting, but
we are not waiting for the LB to go back to ACTIVE. On fast test
hosts this can lead to an immutable error on the shared parent objects.
This patch adds a "wait for active" after the object "wait for deleted".

Change-Id: Ie4c3ffeb194a779b2e4e04ef423d23e3891ad3ce