Fix the scenario parent job to use the live base
The dsvm scenario jobs was using "octavia-dsvm-base" parent job and
not the required "octavia-dsvm-live-base" job. This patch corrects that.

It also changes the "octavia-v2-dsvm-scenario-centos.7" job name to
"octavia-v2-dsvm-scenario-centos-7" to comply with the naming guideline[1].

The same change is applied to the bionic job, which becomes "-bionic".

It also increases the job timeout for "live" jobs to 9,000 seconds, or
2.5 hours to accomidate tcg gate hosts.

[1] https://docs.openstack.org/infra/manual/ \
    drivers.html#consistent-naming-for-jobs-with-zuul-v3

Change-Id: I7aaa86bf0a626cbef17611445d4499969741534d