Fix dependency on requests library version
The code was calling property Response.next which is only available in
requests >= 2.15.0 [1]. At present minimum version of requests library
is 2.14.2 in requirements.txt. Instead of bumping its version (impacts
packaging specially to already released OpenStack versions), we can
solve it on our side with this patch.

The error was:
    AttributeError: 'Response' object has no attribute 'next'
    
[1] https://github.com/requests/requests/blob/v2.15.0/HISTORY.rst

Story: 2004641
Task: 28591

Change-Id: I15b496c740b0c7c8970501d08497ba43cbffda2b