Add iptables-based active/standby scenario test
This patch adds a tempest scenario test for active/standby topology.

This scenario takes a similar approach to scenario proposed in Change-Id
Ibcd5552a67cea650edc72bfaa986357267ca2407 with the difference that it
does not rely on amphora stats API. Instead, it uses iptables to log VIP
traffic and make assertions based on logging.

Co-Authored-By: Michael Johnson <johnsomor@gmail.com>

Change-Id: I24a6fc3387166ec6cdbd57a5ca9f63743748ec68