Update the live jobs to set higher retries
Previously the retry default in Octavia was set very high, 300 retries,
to accomidate OpenStack clouds running without nested virtualization.
In these environments it can take nova up to 20 minutes to boot a VM.

The main Octavia is setting these to a lower, more production friendly
default values. Since some of the OpenStack nodepool instances run on
clouds that do not have nested virtualization enabled, we need to set
these back to 300 for the zuul jobs.

Change-Id: I4ae18e160ecad420095c9650d87e4e3629ee1c0c