Set irrelevant files on CI jobs
All jobs are triggered regardless of files changed, except for files
already defined in octavia-dsvm-base. This means many jobs that are not
impacted by a given change will be run in CI. Not only this is a waste
of CI resources but undeeded jobs may also fail (e.g. timeouts), thus
blocking changes from merging.

This patch defines a more granular set of irrelevant files per job. For
example, it prevents active-standby jobs from running when a change is
limited to normal scenario files.

Change-Id: I9eba013fb0b87afc44187e259b594f9b93bcc5d9