Fix VIP IPv6 subnet discovery order
Code was assuming the subnet existed already and only after creating one
if not. On environments where the subnet does not exist, it raises an
IndexError exception as there are no elements in priv_ipv6_subnet. This
patch fixes the order of the code to first create the subnet if ones
does not exist yet.

Story: 2006280
Task: 35982

Change-Id: I9fa9d7e16f32f516592bf155045d4f4e6a15ed6c