Fix positional formatting and add skip check
This patch fixes an offset positonal argument in a string format that
was throwing an IndexError exception. It also adds a skip check to check
if the required amphora SSH key file is readable.

Additionally, the iptables-based active-standby test has now a better
amphora_ssh_key default value to match with the default path set by the
Octavia Devstack plugin.

Story: 2006259
Task: 35876

Change-Id: Ic887c726dfe151ea582a1bd935ae0b30582f14ea