Set connection_retry_interval back up to 1200
Currently the connection_max_retries setting is set to 300 or 5
minutes. We know that some test nodepool instances can take up to
18 minutes for a VM to boot.
This patch increases the connection_max_retries timeout back up to
1200 or 20 minutes to accomidate those slow hosts.
This will avoid unnecessary job failures[1].
It also increases the check_timeout from 60 seconds to 180 and drops
the check_interval down to 1 second.

[1] https://logs.opendev.org/37/673337/12/check/octavia-v2-dsvm-tls-barbican/ \
    b512ea7/controller/logs/screen-o-cw.txt.gz#_Aug_10_02_43_54_823712

Change-Id: Ib5635d78c97f4ea1d91af588ea582ba5cb20de56