Fix the health monitor traffic test member ERROR
Neutron may be slow to start passing traffic after a health
monitor has been defined on a pool. We have seen "Layer4 timeout"
errors[1] in some of the gate job runs where a few seconds later the
health monitor traffic starts getting a response from the backend
member server.

This patch changes the waiter to allow an "ERROR" status for the
initial member check after the health monitor is added. The waiter
will still timeout if it does not become "ONLINE" as expected.

As the zuul log viewer is broken I can't link to the log line, but it
is at time: Aug 30 01:20:40

[1] https://openstack.fortnebula.com:13808/v1/ \
    AUTH_e8fd161dc34c421a979a9e6421f823e9/logs_58/679358/2/check/ \
    octavia-v2-dsvm-scenario/1bcb675/controller/logs/ \
    octavia-amphora_log.txt.gz

Change-Id: Ic55fabe94627b21a6f347e5822893a8b63cd1afb