Enable fail-fast on the gate queue
Zuul added fail-fast as a new option for project pipelines.
With this set, if any one of the voting jobs fails, zuul
will report a build failure immediately and abort the remaining
jobs.

This patch enables fail-fast on the gate pipeline as any one
failing job means the patch will not merge.

Change-Id: I775dab84b4f1ba7a641ebbed54b6052333502d9c