Add a mixed HTTP and HTTPS scenario test
This patch adds a scenario test covering the "Deploy HTTP and TLS-terminated
HTTPS load balancing on the same IP and backend" usecase from the Octavia
"Basic Load Balancing Cookbook".
It creates a load balancer with two listeners, one HTTPS and one HTTP, that
point to the same backend pool of members. It then checks that the members
are balanced via both listeners.

Change-Id: Ic80eaa10645466ccaffbb19784be6efabeb92aed