Fix test pool to prevent immutability state errors
When a test method completes it must assert its resources are in a
mutable state or following test methods in same test class can fail to
create, update and delete shared test class resources.

This problem has been recently seen in upstream CI [1]. The test method
was not waiting for the load balancer to come back up after a CUD
operation in its child resources, resulting in the next test method
execution failing to create a pool on the test class-level load
balancer.

[1] https://zuul.opendev.org/t/openstack/build/ \
    5d8eb87364c84480b54c87fd6966dbf4/log/controller/logs/tempest_log.txt#7968
    
Story: 2007187
Task: 38313

Change-Id: I487b1fc7e3de8724b16b6f8eabad5e69f5009758