Switch octavia-two-node nodeset to Bionic
Ubuntu Xenial is no longer supported in Devstack. The two-node CI job
fails consistently with:

"WARNING: this script has not been tested on xenial"
"If you wish to run this script anyway run with FORCE=yes"

It is time to update to Ubuntu Bionic.

This patch also removes octavia-v2-dsvm-scenario-ubuntu-bionic from the
check queue as Bionic is now the default Ubuntu version being tested.

Change-Id: Ica2f3289b14b8cdb5437ea3e92bb07a19028e053