Fix amphora list-show to run in a busy environment
If the environment is busy, other amps may be moving around in transient
states, others may be created or deleted in the mid of the test, and
could throw off this test. For example, running tempest tests
concurrently will cause this test to fail occasionally.

Additionally, the amphora list-show test was moved the API test suite.

Task: 38974
Story: 2007389

Co-Authored-By: Adam Harwell <flux.adam@gmail.com>
Change-Id: Iedc7d285dd645e7e80582f25fa90637d2465f21c