Add skip_if_not_implemented to the service client
This patch adds a decorator that wraps the service client methods adding
the ability to raise a skipException if a call to the API returns a
NotImplementedError or UnsupportedOptionError exception.

This is useful for running the tests against provider drivers that do
not support all of the Octavia API features.

Change-Id: I17d4be65130fadf97d6170d22fb07c72672b7573