[ussuri][goal] Updates for python 2.7 drop
OpenStack is dropping the py2.7 support in ussuri cycle except
swift and sorlets which is exception approved in this goal.

Somehow this repo is left to remove the py2.7 requirement
and setup.cfg file cleanup.

Ussuri Communtiy-wide goal:
https://governance.openstack.org/tc/goals/selected/ussuri/drop-py27.html

Also drop the stable/queens|rocky py2 jobs because octavia-tempest-plugin master
code cannot test them because of py2 drop. They are also Extended Maintainance
state so it is ok to not test them on master. Same way Tempest and other
plugins did
- https://review.opendev.org/#/c/722183/

Change-Id: I9b20229a25dc9752e4a813709f0b67bbd104faa5