Fix TLS client auth tests for older amphora
Older amphora have a verison of HAProxy that does not support TLS1.3.
This means that the error returned when the client authentication
certificate is required, but improper is different between the versions.

This patch makes the test more generic to pass when the proper
exception is raised, but will no longer validate the error string
contents as this string varies across the protocol versions used.

Change-Id: Ic08135fdf5fb2e8cf35852bf065a885327a852fa