Fix availability zone API tests
Currently, if you run the availability zone API tests when not using
the no-op drivers, the API tests will fail as they had hard-coded
availability zone names that do not exist.
Since creating availability zones requires reconfiguration of the nova
host deployments, it is not practical to create nova availability zones
for these tests.

This patch adds tempest configuration settings for the required
availability zones, and if not configured will skip the tests that
require the availability zones be present.

Change-Id: I53d5d33ae9a181b79e1d971d56452eeee5dd7759