Change pool create scenario test to wait for operating status
The API test for pool creation has a check to wait for the
operating status of a pool to transition to ONLINE if a
listener has been provided.

The scenario test for pool creation doesn't wait for any
operating status transition after a pool is created with
a listener, and instead always assumes it's OFFLINE.

They should both wait for a transition, since otherwise
drivers that transition to ONLINE quickly, like OVN, can
race and fail the scenario test.

Depends-on: https://review.opendev.org/c/openstack/octavia/+/763530

Change-Id: I9eb3c7ce8af879cd910b06234c99b53670907e49
Story: 2008009
Task: 40659