Build the Go test server on devstack test-config
Every time we update the Go test server code, we have to re-compile it
locally and commit the binary file. Git is not great for binary blobs.
One other reason that carrying the binary in-tree is suboptimal is it
is compiled against amd64 architecture, thus not much multi-arch
friendly.

This change drops the need to compile and carry the test server binary
file in the Git tree. The test server will now be compiled at devstack
test-config phase. On my work station, it takes less than 6 seconds.

Change-Id: I3d8ad18ddd1f23720618d733a9d8a53a6b4e6dfc