Reduce the number of load balancers in PoolAPITest
PoolAPITest creates a new LB in each of its test_*_pool_list functions
(~20 functions), those creations are really time consumming.
Now each test ensures that the pools are deleted on test cleanup, and
that all the test functions use the same LB that has been created in
resource_setup. It speeds up the execution of the test class (from 55min
to 35min in my dev env).

Change-Id: Idb2c45cda4a407c41d314f195af03b4c6a22da07