Update service client access in tempest tests
All the load balancer service clients are registered via the plugin interface[1],
that way Tempest register and create the lazy initialization of registered clients
so that they can be access from there in consistent way.

But octavia-tempest-client create a separate instance of those and access instead of
accessing the registered service client in Tempest. This commit makes all the service clients
access from Tempest registry and remove the separate objects.

[1] https://github.com/openstack/octavia-tempest-plugin/blob/cac3eefc44afca458692e9dadb539c0cb17dc9e4/octavia_tempest_plugin/plugin.py#L54

Change-Id: Ie24909b49baf2c6a886e2ff711e641e36ffe6b50