Increase job timeout for noop-api tests
noop-api tests have been frequently timing out since SQLAlchemy
1.4 release.
An analysis showed that some tests are 55% slower with 1.4, but we
haven't fixed it or improved the code yet. To unblock the gates, set the
timeout to 10800 sec (3h00) for those jobs. We might revert this commit
when we figure out how to improve test exec times.

Change-Id: I68a5357f58b74f0f759cab0bb89a517c384d05e0