Fix request interval in _wait_for_lb_functional
The request_interval parameter is passed to the _wait_for_lb_functional
function but it wasn't used in the function.
Now the function sleeps for request_interval sec for each iteration of
the loop.
It mitigates some random failures with the TCP listener +
least_connection lb_algorithm tests that require a specific delay
between the requests [0]

[0] https://opendev.org/openstack/octavia-tempest-plugin/src/branch/master/octavia_tempest_plugin/tests/scenario/v2/test_traffic_ops.py#L854

Change-Id: Idf0a9fb71178abea732748f92f01eed5a55a3228