Use the same ssl module httpx uses
Depending on what other tempest plugins are installed
the eventlet might've been imported by the time we instantiate
a httpx.Client as part of the test.
Since dnspython 2.2.0, importing eventlet or any part of it effectively
instantiates a dummy httpx.Client instance too, thus pinning the ssl
implementation in httpx to the eventlet's one.
This leads to error in the Client() call of the test, as the ssl lib in
the test module is different from ssl lib in httpx._config,
which fails isinstance check for ssl.SSLContext.

Use the ssl module that is actually currently used by httpx to instantiate
the SSL context to be used with httpx.Client.

Story: 2010902
Task: 48716
Change-Id: I72d80a9981c465723b7fa9aad426118e243b7129