Remove stable/2023.1 job from master gate
stable/2023.1 is planned to be in unmaintained state[1]
and with python 3.8 drop from Tempest[2], we have to pin the
Tempest for stable/2023.1 testing[3]. This means Tempest and
its plugin master does not need to support stable/2023.1.

With that we can remove the stable/2023.1 jobs from tempest
plugin master jobs.

Also disable the doc job currently broken.

[1] https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/message/UDQAC7SR5JAQJE5WBAG54A2MTBVBTJ44/
[2] https://review.opendev.org/c/openstack/tempest/+/933398
[3] https://review.opendev.org/c/openstack/devstack/+/933704

Change-Id: I56043348a15b7c06c9662a200e452d5b2c03f618