Update hacking
This patch updates the version of hacking to be consistent with octavia.
It is also required for flake8 to understand f-strings correctly.
Along with this version update, the patch fixes issues in the code found
by the new versions.

This patch also removes O349 check for importing mock since flake8 has
added H216 which does the same check.

Change-Id: I8e49fc19faff59b2811a7ee09b7ccaa286f73d82